TOS060 2.00 & TOS Drivers 1.00
Fragments of flash.tos, flash.tos/tos and flash.tos/nonfree are kept as
is.

+++ flash.tos/drivers/bdos/fsdrive.c
@@ -281,7 +281,7 @@ static short log2(short n)
     short i;
 
     for (i = 0; n ; i++) {
-        n >>= 1;
+        (unsigned short)n >>= 1;
     }
 
     return(i-1);
